fix(ci): inject generic host runtime capabilities
