+++ moonmind/omnigent/host_runtime.py
@@ -12,8 +12,6 @@ composition root (:mod:`moonmind.omnigent.production`).
 from __future__ import annotations
 
 import hashlib
-import os
-from collections.abc import Mapping
 from dataclasses import dataclass
 from typing import Any, Awaitable, Callable
 
@@ -34,16 +32,10 @@ from moonmind.omnigent.host_ports import (
     OmnigentHostLauncherPort,
     OmnigentHostRegistrationPort,
     OmnigentMountedToolPort,
+    OmnigentRuntimeEnvironmentPort,
     OmnigentSkillDeliveryPort,
     OmnigentWorkspaceMaterializationPort,
     host_correlation_identity,
-)
-from moonmind.omnigent.workspace_intent import authored_required_capabilities
-from moonmind.security.execution_fanout_capabilities import (
-    EXECUTION_FANOUT_REQUIRED_CAPABILITY,
-    ExecutionFanoutCapabilityError,
-    mint_execution_fanout_capability,
-    require_execution_fanout_authorization,
 )
 from moonmind.schemas.agent_runtime_models import AgentExecutionRequest
 
@@ -84,6 +76,7 @@ class GenericOmnigentHostRuntime:
         tool_service: OmnigentMountedToolPort,
         github_credential_service: OmnigentGithubCredentialPort,
         egress_service: OmnigentEgressAttestationPort,
+        runtime_environment_service: OmnigentRuntimeEnvironmentPort,
         registration_waiter: OmnigentHostRegistrationPort,
         host_attestor: OmnigentHostAttestationPort,
         cleanup_service: OmnigentHostCleanupPort,
@@ -95,6 +88,7 @@ class GenericOmnigentHostRuntime:
             tool_service,
             github_credential_service,
             egress_service,
+            runtime_environment_service,
             registration_waiter,
             host_attestor,
             cleanup_service,
@@ -110,90 +104,10 @@ class GenericOmnigentHostRuntime:
         self._tools = tool_service
         self._github_credentials = github_credential_service
         self._egress = egress_service
+        self._runtime_environment = runtime_environment_service
         self._registration = registration_waiter
         self._attestor = host_attestor
         self._cleanup = cleanup_service
-
-    @staticmethod
-    def _execution_fanout_authorization(
-        request: AgentExecutionRequest,
-    ) -> Mapping[str, Any] | None:
-        step_execution = request.step_execution
-        if step_execution is None:
-            return None
-        policy = step_execution.skill_source_policy
-        if "executionFanout" not in policy:
-            return None
-        evidence = policy.get("executionFanout")
-        if not isinstance(evidence, Mapping):
-            raise HarnessPlatformError(
-                "execution fan-out authorization evidence is malformed",
-                code="authorization_denied",
-            )
-        return evidence
-
-    @classmethod
-    def _runtime_environment(
-        cls,
-        *,
-        request: AgentExecutionRequest,
-        plan: OmnigentExecutionPlanEnvelope,
-        host_lease_ref: str,
-        launch_policy: LaunchPolicy,
-    ) -> dict[str, str]:
-        """Mint only the execution capabilities authorized for this host lease."""
-
-        required_capabilities = authored_required_capabilities(request)
-        if EXECUTION_FANOUT_REQUIRED_CAPABILITY not in required_capabilities:
-            return {}
-        try:
-            require_execution_fanout_authorization(
-                required_capabilities,
-                cls._execution_fanout_authorization(request),
-            )
-        except ExecutionFanoutCapabilityError as exc:
-            raise HarnessPlatformError(str(exc), code="authorization_denied") from exc
-        step_execution = request.step_execution
-        workflow_id = (
-            str(step_execution.workflow_id or "").strip()
-            if step_execution is not None
-            else str(request.correlation_id or "").strip()
-        )
-        step_execution_id = (
-            str(step_execution.step_execution_id or "").strip()
-            if step_execution is not None
-            else str(request.correlation_id or "").strip()
-        )
-        runtime_id = str(plan.payload.harnessId or "").strip()
-        moonmind_url = str(
-            os.environ.get("MOONMIND_URL") or "http://api:8000"
-        ).strip()
-        if not workflow_id or not step_execution_id or not runtime_id or not moonmind_url:
-            raise HarnessPlatformError(
-                "execution fan-out runtime identity is incomplete",
-                code=HarnessPlatformFailure.OMNIGENT_HOST_LAUNCH_FAILED,
-            )
-        from moonmind.config.settings import settings
-
-        return {
-            "MOONMIND_URL": moonmind_url,
-            "MOONMIND_AGENT_RUN_ID": step_execution_id,
-            "MOONMIND_TASK_WORKFLOW_ID": workflow_id,
-            "MOONMIND_STEP_ID": step_execution_id,
-            "MOONMIND_RUNTIME_ID": runtime_id,
-            "MOONMIND_EXECUTION_FANOUT_BEARER_TOKEN": (
-                mint_execution_fanout_capability(
-                    secret=str(settings.security.JWT_SECRET_KEY or ""),
-                    parent_workflow_id=workflow_id,
-                    agent_run_id=step_execution_id,
-                    step_id=step_execution_id,
-                    session_id=host_lease_ref,
-                    runtime_id=runtime_id,
-                    source_kind="omnigent",
-                    lifetime_seconds=int(launch_policy.limits["timeoutSeconds"]),
-                )
-            ),
-        }
 
     async def prepare(
         self,
@@ -326,11 +240,13 @@ class GenericOmnigentHostRuntime:
                 prepared.egress_attestation["appliedRuleDigest"]
             ),
         }
-        runtime_environment = self._runtime_environment(
-            request=request,
-            plan=plan,
-            host_lease_ref=host_lease_ref,
-            launch_policy=launch_policy,
+        runtime_environment = dict(
+            self._runtime_environment.build(
+                request=request,
+                plan=plan,
+                host_lease_ref=host_lease_ref,
+                launch_policy=launch_policy,
+            )
         )
         spec = HostLaunchSpec.model_validate(
             {
